// Client/src/app/_models/Page.ts
export type Page = {
    pageSize: number;
    totalRecords: number;
    totalPages: number;
    pageNumber: number;
    firstPage: string;
    lastPage: string;
    previousPage: string;
    nextPage: string;

    minAge: number;
    maxAge: number;
    gender: number;
    
    orderby: number;

  };
  